活動を一覧画面からまとめて削除しようとするとundefinedエラーが生じる不具合の修正

+++ modules/Calendar/actions/MassDelete.php
@@ -22,7 +22,7 @@ class Calendar_MassDelete_Action extends Vtiger_MassDelete_Action {
         }
 		$cvId = $request->get('viewname');
 		foreach($recordIds as $recordId) {
-			if(Users_Privileges_Model::isPermitted($moduleName, 'Delete', $recordId)) {
+			if(isRecordExists($recordId) && Users_Privileges_Model::isPermitted($moduleName, 'Delete', $recordId)) {
 				$recordModel = Vtiger_Record_Model::getInstanceById($recordId, $moduleModel);
 				$parentRecurringId = $recordModel->getParentRecurringRecord();
 				$adb->pquery('DELETE FROM vtiger_activity_recurring_info WHERE activityid=? AND recurrenceid=?', array($parentRecurringId, $recordId));
